# openapi-user.yaml
openapi: 3.0.0
info:
  title: MCCS Alpha User API
  description: |
    ## Introduction

    The MCCS Alpha User API v1 exposes all of the end user functionality currently available in the [alpha version of the MCCS API](https://github.com/technoshantoms/mccs-alpha-api).

    By providing an API, developers who want to create their own front-end user interface for MCCS will have significant flexibility to implement it in whatever way they choose. This means developers can present MCCS in any language, setup their own signup flow, optimize it for whatever devices their users prefer, develop a mobile app, integrate other services such as chat, etc.

    Importantly, an API enables developers to integrate MCCS functionality directly into their own apps (e.g., import transfer data into an accounting application, instruct mutual credit transfers from an e-wallet application, interact with MCCS data via a chat bot, etc.).

    ## Background

    To understand how the MCCS Alpha API works, please read the following design documents:

    - [MCCS Alpha Data Model](https://github.com/technoshantoms/mccs/blob/master/alpha-data-model.md)
    - [MCCS Alpha Functionality](https://github.com/technoshantoms/mccs/blob/master/alpha-functionality.md)

    ## Test Server

...

    ## License
  license:
    name: MIT License
    url: https://github.com/technoshantoms/mccs-alpha-api/blob/dev/LICENSE
  version: '1'
servers:
  - url: http://localhost:8080/api/v1
...
paths:
...
  /transfers:
    post:
      tags:
        - Transfer Credits
      summary: Initiate a transfer
      description: |
        A user can initiate a transfer out of or into the account of its entity, which must then be approved or rejected by the user operating the receiving entity, whose account will be credited or debited accordingly. Both entities must have `tradingAccepted` status in order to set up a transfer between them.

        If the `transfer` parameter is set to `out`, the initiator will create a transfer that will debit funds from the initiator's entity's account. If `transfer` is `in`, the initiator will create a transfer that results in funds being credited to the initiator's entity's account. Either way, the transfer must be approved by the receiver (see `PATCH /transfers/{transferID}`) in order for the inbound or outbound transfer to move to or from the receiver's entity's account.
      requestBody:
        $ref: '#/components/requestBodies/initiateTransfer'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TransferInitiated'
              example:
                id: 1ZceiVuQyGqeUYlC6UIKgEnaBkD
                from: "7132460355005184"
                to: "0382855564717143"
                amount: 177.5
                description: Payment of your invoice number 12345
                status: transferInitiated
                dateProposed: "2020-05-05T14:09:17.446965528Z"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        403:
          $ref: '#/components/responses/Forbidden'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        - jwt: []
    get:
      tags:
        - Review Transfer Activity
      summary: Get a list of transfers
      description: |
        A user can request a list of mutual credit transfers for the account of the entity. Transfers can be filtered by `status` (`all`, `initiated`, `completed` or `cancelled`).

        The `querying_entity_id` is the ID of the entity whose account the information is being requested for. The user requesting must be associated with that entity or no information will be returned.
      parameters:
        - $ref: '#/components/parameters/transferStatus'
        - $ref: '#/components/parameters/queryingEntityIDRequired'
        - $ref: '#/components/parameters/page'
        - $ref: '#/components/parameters/pageSize'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      $ref: '#/components/schemas/TransferView'
                  meta:
                    $ref: '#/components/schemas/Meta'
              examples:
                completed:
                  value:
                    data:
                      - id: 1UZ7G7qJrIlwpVK9iSPXgx0A2xN
                        transfer: out
                        isInitiator: true
                        accountNumber: "1234567887654321"
                        entityName: Rhynyx
                        amount: 1.1,
                        description: Payment of invoice number 12345
                        status: transferCompleted
                        dateProposed: "2019-12-25T12:12:12.123Z"
                        dateCompleted: "2019-12-26T13:13:13.456Z"
                    meta:
                      numberOfResults: 1
                      totalPages: 1
                initiated:
                  value:
                    data:
                      - id: 1UZ7G7qJrIlwpVK9iSPXgx0A2xN
                        transfer: out
                        isInitiator: true
                        accountNumber: "1234567887654321"
                        entityName: Rhynyx
                        amount: 1.2,
                        description: Payment of invoice number 12345
                        status: transferInitiated
                        dateProposed: "2019-12-25T12:11:11.456Z"
                    meta:
                      numberOfResults: 1
                      totalPages: 1
                cancelled:
                  value:
                    data:
                      - id: 1UZ7G7qJrIlwpVK9iSPXgx0A2xN
                        transfer: out
                        isInitiator: true
                        accountNumber: "1234567887654321"
                        entityName: Rhynyx
                        amount: 1.3,
                        description: Payment of invoice number 12345
                        status: transferCancelled
                        cancellationReason: Wrong amount, I will send a corrected request
                        dateProposed: "2019-12-25T12:11:11.456Z"
                    meta:
                      numberOfResults: 1
                      totalPages: 1
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        - jwt: []
  /transfers/{transferID}:
    patch:
      tags:
        - Transfer Credits
      summary: Confirm or cancel a transfer
      description: |
        The receiver can either `accept` or `reject` the transfer by specifying it in the action parameter.

        The initiator of the transfer can `cancel` the transfer before the receiver has accepted or rejected it.

        If a transfer is rejected or cancelled, a `cancellationReason` can be provided so that the other party understands why the initiator or receiver cancelled or rejected the transfer.      
      parameters:
        - $ref: '#/components/parameters/transferID'
      requestBody:
        $ref: '#/components/requestBodies/confirmOrCancelTransfer'
      responses:
        200:
          description: OK
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    $ref: '#/components/schemas/TransferView'
              examples:
                acceptResponse:
                  value:
                    id: "1UZ7G7qJrIlwpVK9iSPXgx0A2xN"
                    transfer: out
                    isInitiator: true
                    accountNumber: "1234567887654321"
                    entityName: Rhynyx
                    amount: 1.1,
                    description: Payment of invoice number 12345
                    status: transferCompleted
                    dateProposed: "2019-12-25T12:12:12.123Z"
                    dateCompleted: "2019-12-26T13:13:13.456Z"
                rejectResponse:
                  value:
                    id: "1UZ7G7qJrIlwpVK9iSPXgx0A2xN"
                    transfer: out
                    isInitiator: true
                    accountNumber: "1234567887654321"
                    entityName: Rhynyx
                    amount: 1.1,
                    description: Payment of invoice number 12345
                    status: transferCancelled
                    cancellationReason: some reason for rejecting
                    dateProposed: "2019-12-25T12:12:12.123Z"
                cancelResponse:
                  value:
                    id: "1UZ7G7qJrIlwpVK9iSPXgx0A2xN"
                    transfer: out
                    isInitiator: true
                    accountNumber: "1234567887654321"
                    entityName: Rhynyx
                    amount: 1.1,
                    description: Payment of invoice number 12345
                    status: transferCancelled
                    cancellationReason: some reason for cancelling
                    dateProposed: "2019-12-25T12:12:12.123Z"
        400:
          $ref: '#/components/responses/BadRequest'
        401:
          $ref: '#/components/responses/Unauthorized'
        429:
          $ref: '#/components/responses/TooManyRequests'
        500: 
          $ref: '#/components/responses/ServerError'
      security:
        - jwt: []
...
components:
  schemas:
...
    TransferInitiated:
      type: object
      title: TransferInitiated
      description: An object, shown only to the initiator of a transfer immediately after creation, representing the confirmation of an initiated transfer of mutual credits from one entity's account to another
      properties:
        id:
          type: string
        from:
          type: string
        to:
          type: string
        amount:
          type: number
        description:
          type: string
        status:
          type: string
          enum:
            - transferInitiated
        dateProposed:
          type: string
    TransferView:
      type: object
      title: TransferView
      description: An object representing an initiated/completed/cancelled transfer of mutuals credits from one entity to another
      properties:
        id:
          type: string
        transfer:
          type: string
          enum:
            - in
            - out
        isInitiator:
          type: boolean
        accountNumber:
          type: string
        entityName:
          type: string
        amount:
          type: number
        description:
          type: string
        status:
          type: string
          enum:
            - transferInitiated
            - transferCompleted
            - transferCancelled
        cancellationReason:
          type: string
        dateProposed:
          type: string
        dateCompleted:
          type: string
...
    Error:
      type: object
      title: Error
      description: An error object
      properties:
        message:
          type: string
    Meta:
      type: object
      properties:
        numberOfResults:
          type: integer
        totalPages:
          type: integer
  parameters:
...
    queryingEntityIDRequired:
      name: querying_entity_id
      description: The entity ID to which the account is linked
      in: query
      required: true
      schema:
        type: string
        example: 5e561916ca06e1c8596eee9e
...
    transferID:
      name: transferID
      description: The unique transfer ID
      in: path
      required: true
      schema:
        type: string
        example: 1UZ7G7qJrIlwpVK9iSPXgx0A2xN
    transferStatus:
      name: status
      description: The status of the transfer
      in: query
      required: true
      schema:
        type: string
        enum:
          - all
          - initiated
          - completed
          - cancelled
    page:
      name: page
      description: The page number
      in: query
      schema:
          type: integer
          default: 1
    pageSize:
      name: page_size
      description: The number of results per page
      in: query
      schema:
        type: integer
        default: 10
        minimum: 1
        maximum: 100
  requestBodies:
...
    initiateTransfer:
      description: Initiate a transfer to/from entity's account from/to another entity
      required: true
      content:
        application/json:
          schema:
            type: object
            properties:
              transfer:
                type: string
                enum:
                  - in
                  - out
              initiator:
                type: string
              receiver:
                type: string
              amount:
                type: number
              description:
                type: string
          example:
            transfer: out
            initiator: "7132460355005184"
            receiver: "1234567887654321"
            amount: 1.1
            description: Payment of invoice number 12345
    confirmOrCancelTransfer:
      required: true
      content:
        application/json:
          schema:
            type: object
            required:
              - action
            properties:
              action:
                type: string
                enum:
                  - accept
                  - reject
                  - cancel
              cancellationReason:
                type: string
          examples:
            accept:
              value:
                action: accept
            reject:
              value:
                action: reject
                cancellationReason: some reason for rejecting
            cancel:
              value:
                action: cancel
                cancellationReason: some reason for cancelling
  responses:
    BadRequest:
      description: The request is missing a required parameter.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
          example:
            errors:
              - message: <named> parameter is missing.
    Unauthorized:
      description: There was an issue with the authentication data for the request.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
          example:
            errors:
              - message: Could not authenticate you.
    Forbidden:
      description: User does not have permission to access the resource.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
          example:
            errors:
              - message: Permission denied.
    TooManyRequests:
      description: The request limit for this resource has been reached for the current rate limit window.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
          example:
            errors:
              - message: Rate limit exceeded.
    ServerError:
      description: An unknown internal error occurred.
      content:
        application/json:
          schema:
            type: object
            properties:
              errors:
                type: array
                items:
                  $ref: '#/components/schemas/Error'
          example:
            errors:
              - message: Internal server error triggered.
  securitySchemes:
    jwt:
      type: http
      scheme: bearer
      bearerFormat: JWT